backend/requirements.txt
flask
flask-cors
google-genai
python-dotenv
tensorflow-cpu
pillow
numpy
gunicorn
openai
tenacity
Flask-Limiter
